Control npm testing with environment variables
The NPM based tests can now skip the stopping of the docker-compose
cluster by defining the SKIP_SHUTDOWN environment variable.

+++ test/run_npm_test.sh
@@ -90,6 +90,9 @@ npm test
 rval=$?
 
 # Stop MariaDB servers
-docker-compose down -v
+if [ -z  "$SKIP_SHUTDOWN" ]
+then
+    docker-compose down -v
+fi
 
 exit $rval
